+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed fallback model name to be used as `model` argument instead of inexistent `model_name` on ReAct agents.
 - Fixed missing `assignee_id` on `RepoClient.update_or_create_merge_request` abstract method.
 
 ## [0.1.0-alpha.8] - 2024-12-11